pyproject.toml
[project]
name = "receipt-automation"
version = "0.1.0"
description = "영수증 매칭 자동화 프로토타입"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]